refactor: Reorganize navigation files by feature and rename MainNavHost function (#52)
- Moved navigation files (Route, Navigation) to respective feature directories:
    - `auth`
    - `bottomtab`
    - `chatroom`
    - `createpost`
    - `productdetail`
- Updated import statements across the application to reflect the new file locations.
- Renamed `TabNavHost` to `MainNavHost` for clarity.
