+++ app/src/main/java/com/example/rentit/navigation/AuthNavHost.kt
@@ -5,6 +5,8 @@ import androidx.annotation.RequiresApi
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
+import com.example.rentit.navigation.auth.AuthRoute
+import com.example.rentit.navigation.auth.authGraph
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
