+++ app/src/main/java/com/example/rentit/navigation/MainNavHost.kt
@@ -8,10 +8,15 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
+import com.example.rentit.navigation.bottomtab.BottomTabRoute
+import com.example.rentit.navigation.bottomtab.bottomTabGraph
+import com.example.rentit.navigation.chatroom.chatRoomGraph
+import com.example.rentit.navigation.createpost.createPostGraph
+import com.example.rentit.navigation.productdetail.productDetailGraph
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
-fun TabNavHost(navHostController: NavHostController, paddingValues: PaddingValues) {
+fun MainNavHost(navHostController: NavHostController, paddingValues: PaddingValues) {
     NavHost(navController = navHostController, startDestination = BottomTabRoute.Home.route, modifier = Modifier.padding(paddingValues)){
         bottomTabGraph(navHostController)
         productDetailGraph(navHostController)
