+++ app/src/main/java/com/example/rentit/navigation/auth/AuthRoute.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.navigation
+package com.example.rentit.navigation.auth
 
 import kotlinx.serialization.Serializable
