+++ app/src/main/java/com/example/rentit/navigation/chatroom/ChatRoomNavigation.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.navigation
+package com.example.rentit.navigation.chatroom
 
 import android.os.Build
 import androidx.annotation.RequiresApi
