+++ app/src/main/java/com/example/rentit/navigation/createpost/CreatePostNavigation.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.navigation
+package com.example.rentit.navigation.createpost
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
