+++ app/src/main/java/com/example/rentit/navigation/createpost/CreatePostRoute.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.navigation
+package com.example.rentit.navigation.createpost
 
 import kotlinx.serialization.Serializable
