+++ app/src/main/java/com/example/rentit/navigation/productdetail/ProductDetailNavigation.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.navigation
+package com.example.rentit.navigation.productdetail
 
 import android.os.Build
 import androidx.annotation.RequiresApi
