+++ app/src/main/java/com/example/rentit/presentation/mypage/MyPageScreen.kt
@@ -56,7 +56,7 @@ import com.example.rentit.common.theme.pretendardFamily
 import com.example.rentit.data.product.dto.ProductDto
 import com.example.rentit.data.user.dto.ReservationDto
 import com.example.rentit.common.component.ProductListItem
-import com.example.rentit.navigation.navigateToProductDetail
+import com.example.rentit.navigation.productdetail.navigateToProductDetail
 import com.example.rentit.presentation.mypage.components.MyRentalHistoryListItem
 
 @RequiresApi(Build.VERSION_CODES.O)
